enlarge bottom/top plate by 2cm
This gives us enough space to, if we have to, connect each individual
PCB via USB to the raspberry pi. It also makes sure that the power
supply has the recommended safety distance of 4mm on all sides. Further,
we should be able to use all the angles in the case now.